Move geocoding functionality into its own class
