+++ gnd/src/main/java/com/google/android/gnd/ui/offlinearea/OfflineAreaListAdapter.java
@@ -16,31 +16,18 @@
 
 package com.google.android.gnd.ui.offlinearea;
 
-import android.location.Address;
-import android.location.Geocoder;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gnd.R;
 import com.google.android.gnd.databinding.OfflineAreasListItemBinding;
 import com.google.android.gnd.model.basemap.OfflineArea;
+import com.google.android.gnd.system.GeocodingManager;
 import com.google.android.gnd.ui.common.Navigator;
 import com.google.common.collect.ImmutableList;
-import java.util.List;
-import java8.util.Optional;
-import timber.log.Timber;
 
 class OfflineAreaListAdapter extends RecyclerView.Adapter<OfflineAreaListAdapter.ViewHolder> {
-
-  static class AddressNotFoundException extends Exception {
-    public AddressNotFoundException(String message) {
-      super(message);
-    }
-  }
-
   private final Navigator navigator;
   private ImmutableList<OfflineArea> offlineAreas;
 
@@ -50,7 +37,7 @@ class OfflineAreaListAdapter extends RecyclerView.Adapter<OfflineAreaListAdapter
     public int position;
     private ImmutableList<OfflineArea> areas;
     private final Navigator navigator;
-    private final Geocoder geocoder;
+    private final GeocodingManager geocoder;
 
     ViewHolder(
         OfflineAreasListItemBinding binding,
@@ -60,7 +47,7 @@ class OfflineAreaListAdapter extends RecyclerView.Adapter<OfflineAreaListAdapter
       this.binding = binding;
       this.areas = areas;
       this.navigator = navigator;
-      this.geocoder = new Geocoder(binding.getRoot().getContext());
+      this.geocoder = new GeocodingManager(binding.getRoot().getContext());
       binding.offlineAreaName.setOnClickListener(this);
     }
 
@@ -91,41 +78,9 @@ class OfflineAreaListAdapter extends RecyclerView.Adapter<OfflineAreaListAdapter
 
   @Override
   public void onBindViewHolder(@NonNull ViewHolder viewHolder, int position) {
-    String areaName =
-        viewHolder.binding.getRoot().getContext().getString(R.string.offline_areas_unknown_area);
-
     OfflineArea area = offlineAreas.get(position);
-    String id = area.getId();
-    LatLng center = area.getBounds().getCenter();
-
-    try {
-      List<Address> addresses =
-          viewHolder.geocoder.getFromLocation(center.latitude, center.longitude, 1);
-
-      if (addresses.isEmpty()) {
-        throw new AddressNotFoundException("No address found for area.");
-      }
-
-      Address address = addresses.get(0);
-
-      areaName =
-          Optional.ofNullable(address.getFeatureName())
-              .or(() -> Optional.ofNullable(address.getLocality()))
-              .or(() -> Optional.ofNullable(address.getSubAdminArea()))
-              .or(() -> Optional.ofNullable(address.getAdminArea()))
-              .or(() -> Optional.ofNullable(address.getCountryName()))
-              .orElse(areaName);
-    } catch (Exception e) {
-      Timber.e(
-          e,
-          "Couldn't get address for area: %s, lat: %f, lng: %f",
-          id,
-          center.latitude,
-          center.longitude);
-    } finally {
-      viewHolder.binding.offlineAreaName.setText(areaName);
-      viewHolder.position = position;
-    }
+    viewHolder.binding.offlineAreaName.setText(viewHolder.geocoder.getOfflineAreaName(area));
+    viewHolder.position = position;
   }
 
   @Override
